backend/requirements.txt


fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
websockets
python-dotenv
assemblyai
asyncpg # Better async driver for postgresql with FastAPI

